Improve music metadata JSON files
This puts the tracks in order, adds records, and
even adds tracks that weren't put in the game.

Eventually, I will add support for scrobbling
tracks that play on records, and a screen to show
all the albums that are currently loaded (which
can include tracks that aren't in the game,
although they will be unplayable).